made MainFields a singleton

+++ TicTacToe/src/main/java/com/organization4242/tictactoe/ai/TicTacToeAI.java
@@ -11,18 +11,18 @@ import java.util.List;
  */
 public class TicTacToeAI {
     public static Byte[] nextMove(MainField state) {
-        int activeField = state.getActiveField();
+        byte activeField = state.getActiveField();
         boolean order = state.getOrder();
         Byte[] move = new Byte[]{0,0};
-        List<List<Integer>> templateField = new ArrayList<List<Integer>>();
-        List<List<Integer>> inspectedField = new ArrayList<List<Integer>>();
+        List<List<Byte>> templateField = new ArrayList<List<Byte>>();
+        List<List<Byte>> inspectedField = new ArrayList<List<Byte>>();
         templateField.addAll(state.getFields().get(activeField));
         for (int i = 0; i < 3; i++) {
             for(int j = 0; j < 3; j++) {
                 inspectedField.addAll(templateField);
                 if(state.getFields().get(activeField).get(i).get(j) == 0) {
-                    inspectedField.get(i).set(j, 1);
-                    if(CheckForWin.Winner(inspectedField) == 1) {
+                    inspectedField.get(i).set((byte) j, (byte) 1);
+                    if(CheckForWin.Winner(inspectedField) == (byte) 1) {
                         move = new Byte[]{(byte) i,(byte) j};
                         return move;
                     }
