+++ TicTacToe/src/main/java/com/organization4242/tictactoe/model/CheckForWin.java
@@ -22,7 +22,7 @@ public class CheckForWin {
         }
         return false;
     }
-    public static Byte Winner(List<List<Integer>> matrix) {
+    public static Byte Winner(List<List<Byte>> matrix) {
         int sums[] = {0,0,0,0,0,0,0,0};
         for (int i = 0; i < 3; i++) {
             for (int j = 0; j < 3; j++) {
